pyproject.toml
[project]
name = "kt-prompt-optimizer"
version = "0.1.0"
description = "Advanced Korean sentence classification prompt optimizer"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "dataclasses-json>=0.6.0",
    "typing-extensions>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemini", "services", "models", "utils"]